+++ lib/image_helper.dart
@@ -1,14 +1,35 @@
+import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
 
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:gif_view/gif_view.dart';
+import 'package:lottie/lottie.dart';
 
-enum ImageType { asset, file, network }
+import 'default_error_widget.dart';
+
+enum ImageType {
+  asset,
+  file,
+  network,
+  svg,
+  networkSvg,
+  memory,
+  gifAsset,
+  gifMemory,
+  GifNetwork,
+  JsonAsset,
+  JsonFile,
+  JsonNetwork,
+  JsonMemory,
+}
 enum ImageShape { circle, rectangle, oval, none }
 
 class ImageHelper extends StatelessWidget {
   /// image type is for provide different image loader with enum type
-  /// it provide image changes as asset, file, network, or cached
+  /// it provide image changes as assetFlag, file, network, or cached
   final ImageType imageType;
 
   /// image shape is for define the shape around the image
@@ -25,14 +46,14 @@ class ImageHelper extends StatelessWidget {
   final double? height, width;
 
   /// change image fit into source, default is null
-  final BoxFit? boxFit;
+  final BoxFit boxFit;
 
   /// error and loader Builders have default value in case of null
   final Widget? errorBuilder, loaderBuilder;
 
   /// default is low
   /// Quality levels for image sampling in [ImageFilter] and [Shader] objects that sample
-  /// images and for [Canvas] operations that render images.
+  /// assetFlag and for [Canvas] operations that render assetFlag.
   ///
   /// When scaling up typically the quality is lowest at [none], higher at [low] and [medium],
   /// and for very large scale factors (over 10x) the highest at [high].
@@ -42,13 +63,13 @@ class ImageHelper extends StatelessWidget {
   /// reductions. Otherwise, [low] and [high] provide similar effects for reductions of
   /// between 50% and 100% but the image may lose detail and have dropouts below 50%.
   ///
-  /// To get high quality when scaling images up and down, or when the scale is
+  /// To get high quality when scaling assetFlag up and down, or when the scale is
   /// unknown, [medium] is typically a good balanced choice.
   ///
   /// ![](https://flutter.github.io/assets-for-api-docs/assets/dart-ui/filter_quality.png)
   ///
   /// When building for the web using the `--web-renderer=html` option, filter
-  /// quality has no effect. All images are rendered using the respective
+  /// quality has no effect. All assetFlag are rendered using the respective
   /// browser's default setting.
   ///
   /// See also:
@@ -105,6 +126,97 @@ class ImageHelper extends StatelessWidget {
   /// {@macro flutter.painting.BoxDecoration.clip}
   final BorderRadiusGeometry? borderRadius;
 
+  /// curve that use to load image with animation, not applied to all type of assetFlag
+  final Curve fadeInAnime, fadeOutAnime;
+
+  /// duration of loading image, not applied to all type of assetFlag
+  final Duration fadeInDuration, fadeOutDuration;
+
+  /// opacity animation for image, default is null
+  final Animation<double>? opacity;
+
+  /// image repeat for repeat image , default is ImageRepeat.noRepeat
+  final ImageRepeat imageRepeat;
+
+  /// Whether to exclude this image from semantics.
+  ///
+  /// Useful for assetFlag which do not contribute meaningful information to an
+  /// application.
+  final bool excludeFromSemantics;
+
+  /// The center slice for a nine-patch image.
+  ///
+  /// The region of the image inside the center slice will be stretched both
+  /// horizontally and vertically to fit the image into its destination. The
+  /// region of the image above and below the center slice will be stretched
+  /// only horizontally and the region of the image to the left and right of
+  /// the center slice will be stretched only vertically.
+  final Rect? centerSlice;
+
+  /// Whether to continue showing the old image (true), or briefly show nothing
+  /// (false), when the image provider changes. The default value is false.
+  ///
+  /// ## Design discussion
+  ///
+  /// ### Why is the default value of [gaplessPlayback] false?
+  ///
+  /// Having the default value of [gaplessPlayback] be false helps prevent
+  /// situations where stale or misleading information might be presented.
+  /// Consider the following case:
+  ///
+  /// We have constructed a 'Person' widget that displays an avatar [Image] of
+  /// the currently loaded person along with their name. We could request for a
+  /// new person to be loaded into the widget at any time. Suppose we have a
+  /// person currently loaded and the widget loads a new person. What happens
+  /// if the [Image] fails to load?
+  ///
+  /// * Option A ([gaplessPlayback] = false): The new person's name is coupled
+  /// with a blank image.
+  ///
+  /// * Option B ([gaplessPlayback] = true): The widget displays the avatar of
+  /// the previous person and the name of the newly loaded person.
+  ///
+  /// This is why the default value is false. Most of the time, when you change
+  /// the image provider you're not just changing the image, you're removing the
+  /// old widget and adding a new one and not expecting them to have any
+  /// relationship. With [gaplessPlayback] on you might accidentally break this
+  /// expectation and re-use the old widget.
+  final bool gaplessPlayback;
+
+  /// Whether to paint the image with anti-aliasing.
+  ///
+  /// Anti-aliasing alleviates the sawtooth artifact when the image is rotated.
+  final bool isAntiAlias;
+
+  /// Whether to paint the image in the direction of the [TextDirection].
+  ///
+  /// If this is true, then in [TextDirection.ltr] contexts, the image will be
+  /// drawn with its origin in the top left (the "normal" painting direction for
+  /// assetFlag); and in [TextDirection.rtl] contexts, the image will be drawn with
+  /// a scaling factor of -1 in the horizontal direction so that the origin is
+  /// in the top right.
+  ///
+  /// This is occasionally used with assetFlag in right-to-left environments, for
+  /// assetFlag that were designed for left-to-right locales. Be careful, when
+  /// using this, to not flip assetFlag with integral shadows, text, or other
+  /// effects that will look incorrect when flipped.
+  ///
+  /// If this is true, there must be an ambient [Directionality] widget in
+  /// scope.
+  final bool matchTextDirection;
+
+  /// A Semantic description of the image.
+  ///
+  /// Used to provide a description of the image to TalkBack on Android, and
+  /// VoiceOver on iOS.
+  final String? semanticLabel;
+
+  /// used for gif image frame rate of image loading
+  final int? frameRate;
+
+  /// box border surounded image
+  final BoxBorder? boxBorder;
+
   ImageHelper(
       {required this.image,
       required this.imageType,
@@ -112,7 +224,7 @@ class ImageHelper extends StatelessWidget {
       this.color,
       this.height,
       this.width,
-      this.boxFit,
+      this.boxFit = BoxFit.contain,
       this.errorBuilder,
       this.filterQuality = FilterQuality.low,
       this.loaderBuilder,
@@ -121,7 +233,21 @@ class ImageHelper extends StatelessWidget {
       this.scale = 1.0,
       this.borderRadius,
       this.defaultLoaderColor,
-      this.defaultErrorBuilderColor});
+      this.defaultErrorBuilderColor,
+      this.fadeInAnime = Curves.easeIn,
+      this.fadeOutAnime = Curves.easeOut,
+      this.fadeInDuration = const Duration(milliseconds: 300),
+      this.fadeOutDuration = const Duration(milliseconds: 300),
+      this.opacity,
+      this.imageRepeat = ImageRepeat.noRepeat,
+      this.excludeFromSemantics = false,
+      this.centerSlice,
+      this.gaplessPlayback = false,
+      this.isAntiAlias = false,
+      this.matchTextDirection = false,
+      this.semanticLabel,
+      this.frameRate = 15,
+      this.boxBorder});
 
   @override
   Widget build(BuildContext context) {
@@ -138,15 +264,25 @@ class ImageHelper extends StatelessWidget {
   }
 
   Widget get _rounded => Container(
-        child: _loadImage,
         clipBehavior: Clip.antiAlias,
-        decoration: BoxDecoration(borderRadius: borderRadius),
+        decoration:
+            BoxDecoration(borderRadius: borderRadius, border: boxBorder),
+        child: Container(
+            child: _loadImage,
+            clipBehavior: Clip.hardEdge,
+            decoration: BoxDecoration(borderRadius: borderRadius)),
       );
 
   Widget get _circle => Container(
-        child: _loadImage,
-        clipBehavior: Clip.antiAlias,
-        decoration: BoxDecoration(shape: BoxShape.circle),
+    clipBehavior: Clip.antiAlias,
+        decoration: BoxDecoration(
+          shape: BoxShape.circle,
+        ),
+        child: Container(
+          child: _loadImage,
+          clipBehavior: Clip.antiAlias,
+          decoration: BoxDecoration(shape: BoxShape.circle, border: boxBorder),
+        ),
       );
 
   Widget get _oval => ClipOval(
@@ -162,33 +298,71 @@ class ImageHelper extends StatelessWidget {
         return _file;
       case ImageType.network:
         return _cached;
+      case ImageType.svg:
+        return _svgLocalIcon;
+      case ImageType.networkSvg:
+        return _svgNetworkIcon;
+      case ImageType.memory:
+        return memoryImage;
+      case ImageType.gifAsset:
+        return _gifAsset;
+      case ImageType.gifMemory:
+        return _gifMemory;
+      case ImageType.GifNetwork:
+        return _gifNetwork;
+      case ImageType.JsonAsset:
+        return _jsonAsset;
+      case ImageType.JsonFile:
+        return _jsonFile;
+      case ImageType.JsonNetwork:
+        return _jsonNetwork;
+      case ImageType.JsonMemory:
+        return _jsonMemory;
     }
   }
 
-  Widget get _file => Image.file(File(image),
-      color: color,
-      cacheHeight: height!.toInt(),
-      cacheWidth: width!.toInt(),
-      height: height,
-      width: width,
-      fit: boxFit,
-      errorBuilder: (context, error, stackTrace) => _errorBuilder,
-      filterQuality: filterQuality,
-      colorBlendMode: blendMode,
-      alignment: alignment,
-      scale: scale);
+  Widget get _file => Image.file(
+        File(image),
+        color: color,
+        cacheHeight: height!.toInt(),
+        cacheWidth: width!.toInt(),
+        height: height,
+        width: width,
+        fit: boxFit,
+        errorBuilder: (context, error, stackTrace) => _errorBuilder,
+        filterQuality: filterQuality,
+        colorBlendMode: blendMode,
+        alignment: alignment,
+        scale: scale,
+        opacity: opacity,
+        repeat: imageRepeat,
+        excludeFromSemantics: excludeFromSemantics,
+        centerSlice: centerSlice,
+        gaplessPlayback: gaplessPlayback,
+        isAntiAlias: false,
+        matchTextDirection: matchTextDirection,
+        semanticLabel: semanticLabel,
+      );
 
   Widget get _cached => CachedNetworkImage(
         imageUrl: image,
         color: color,
+        useOldImageOnUrlChange: false,
+        placeholderFadeInDuration: fadeInDuration,
         colorBlendMode: blendMode,
         filterQuality: filterQuality,
         height: height,
         width: width,
         fit: boxFit,
+        fadeInCurve: Curves.fastOutSlowIn,
+        fadeInDuration: Duration(milliseconds: 300),
         progressIndicatorBuilder: (context, url, progress) => _loaderBuilder,
         errorWidget: (context, url, error) => _errorBuilder,
         alignment: alignment,
+        matchTextDirection: matchTextDirection,
+        repeat: imageRepeat,
+        fadeOutCurve: fadeOutAnime,
+        fadeOutDuration: fadeOutDuration,
       );
 
   Widget get _asset => Image.asset(
@@ -202,7 +376,153 @@ class ImageHelper extends StatelessWidget {
         color: color,
         scale: scale,
         errorBuilder: (context, error, stackTrace) => _errorBuilder,
+        repeat: imageRepeat,
+        matchTextDirection: matchTextDirection,
+        isAntiAlias: isAntiAlias,
+        gaplessPlayback: gaplessPlayback,
+        centerSlice: centerSlice,
+        excludeFromSemantics: excludeFromSemantics,
+        opacity: opacity,
+        semanticLabel: semanticLabel,
       );
+
+  Widget get _svgLocalIcon => SvgPicture.asset(
+        image,
+        height: height,
+        width: width,
+        color: color,
+        colorBlendMode: blendMode,
+        fit: boxFit,
+        alignment: alignment,
+        placeholderBuilder: (context) => _errorBuilder,
+        excludeFromSemantics: excludeFromSemantics,
+        matchTextDirection: matchTextDirection,
+        semanticsLabel: semanticLabel,
+      );
+
+  Widget get _svgNetworkIcon => SvgPicture.network(
+        image,
+        height: height,
+        width: width,
+        color: color,
+        colorBlendMode: blendMode,
+        fit: boxFit,
+        alignment: alignment,
+        placeholderBuilder: (context) => _loaderBuilder,
+        semanticsLabel: semanticLabel,
+        matchTextDirection: matchTextDirection,
+        excludeFromSemantics: excludeFromSemantics,
+      );
+
+  Widget get memoryImage => Image.memory(
+        dataFromBase64String(image),
+        width: 100,
+        height: 100,
+        color: color,
+        scale: scale,
+        fit: boxFit,
+        alignment: alignment,
+        filterQuality: filterQuality,
+        errorBuilder: (context, error, stackTrace) => _errorBuilder,
+        colorBlendMode: blendMode,
+        excludeFromSemantics: excludeFromSemantics,
+        matchTextDirection: matchTextDirection,
+        semanticLabel: semanticLabel,
+        opacity: opacity,
+        centerSlice: centerSlice,
+        gaplessPlayback: gaplessPlayback,
+        isAntiAlias: isAntiAlias,
+        repeat: imageRepeat,
+      );
+
+  Widget get _gifMemory => GifView.memory(
+        Uint8List.fromList(_gifCodeUnite),
+        width: width,
+        height: height,
+        color: color,
+        centerSlice: centerSlice,
+        matchTextDirection: matchTextDirection,
+        fit: boxFit,
+        alignment: alignment,
+        repeat: imageRepeat,
+        colorBlendMode: blendMode,
+        filterQuality: filterQuality,
+        progress: _loaderBuilder,
+        isAntiAlias: isAntiAlias,
+        frameRate: frameRate ?? 15,
+      );
+
+  Widget get _gifAsset => GifView.asset(
+        image,
+        width: width,
+        height: height,
+        color: color,
+        centerSlice: centerSlice,
+        matchTextDirection: matchTextDirection,
+        fit: boxFit,
+        alignment: alignment,
+        repeat: imageRepeat,
+        colorBlendMode: blendMode,
+        filterQuality: filterQuality,
+        progress: _loaderBuilder,
+        isAntiAlias: isAntiAlias,
+        frameRate: frameRate ?? 15,
+      );
+
+  Widget get _gifNetwork => GifView.network(
+        image,
+        width: width,
+        height: height,
+        color: color,
+        centerSlice: centerSlice,
+        matchTextDirection: matchTextDirection,
+        fit: boxFit,
+        alignment: alignment,
+        repeat: imageRepeat,
+        colorBlendMode: blendMode,
+        filterQuality: filterQuality,
+        progress: _loaderBuilder,
+        isAntiAlias: isAntiAlias,
+        frameRate: frameRate ?? 15,
+      );
+
+  Widget get _jsonAsset => Lottie.asset(image,
+      width: width,
+      alignment: alignment,
+      errorBuilder: (context, error, stackTrace) => _errorBuilder,
+      repeat: true,
+      fit: boxFit,
+      height: height);
+
+  Widget get _jsonNetwork => Lottie.network(image,
+      width: width,
+      alignment: alignment,
+      errorBuilder: (context, error, stackTrace) => _errorBuilder,
+      repeat: true,
+      fit: boxFit,
+      height: height);
+
+  Widget get _jsonMemory => Lottie.memory(dataFromBase64String(image),
+      width: width,
+      alignment: alignment,
+      errorBuilder: (context, error, stackTrace) => _errorBuilder,
+      repeat: true,
+      fit: boxFit,
+      height: height);
+
+  Widget get _jsonFile => Lottie.file(image,
+      width: width,
+      alignment: alignment,
+      errorBuilder: (context, error, stackTrace) => _errorBuilder,
+      repeat: true,
+      fit: boxFit,
+      height: height);
+
+  List<int> get _gifCodeUnite => image.codeUnits;
+
+  Uint8List dataFromBase64String(String base64String) {
+    return base64Decode(base64String);
+  }
 
   Widget get _loaderBuilder => Center(
         child: SizedBox(
@@ -215,22 +535,19 @@ class ImageHelper extends StatelessWidget {
         ),
       );
 
-  Widget get _errorBuilder => FittedBox(
-        fit: BoxFit.contain,
-        child: errorBuilder == null ? _defaultErrorBuilder : errorBuilder!,
-      );
-
-  Widget get _defaultErrorBuilder => Icon(
-        Icons.image_not_supported_rounded,
-        color: defaultErrorBuilderColor,
-        size: _imageNotFoundHeight > _imageNotFoundWidth
-            ? _imageNotFoundWidth
-            : _imageNotFoundHeight,
-      );
-
-  double get _imageNotFoundHeight => height != null ? height! / 2 : 30.0;
-
-  double get _imageNotFoundWidth => width != null ? width! / 2 : 30.0;
+  Widget get _errorBuilder => errorBuilder == null
+      ? DefaultErrorWidget(
+          alignment: alignment,
+          boxFit: boxFit,
+          color: color,
+          height: height,
+          semanticLabel: semanticLabel,
+          matchTextDirection: matchTextDirection,
+          excludeFromSemantics: excludeFromSemantics,
+          width: width,
+          blendMode: blendMode,
+        )
+      : errorBuilder!;
 
   Widget get _loader => loaderBuilder == null
       ? CircularProgressIndicator(