+++ CHANGELOG.md
@@ -1,4 +1,64 @@
 ## [1.0.0]
+
 * initial release.
+
 ## [1.0.1]
+
 * fix change asset image color.
+
+## [1.0.2]
+
+* added svg support from asset or network
+
+## [1.0.3]
+
+* fixing svg box fit issue
+
+## [1.0.4]
+
+* added new Future for all image type.
+* also added memory image loading type.
+
+## [1.0.5]
+
+* fixed issue with assets images.
+
+## [1.0.6]
+
+* fixed issue with assets images.
+
+## [1.0.7]
+
+* fixed issue with assets images.
+
+## [1.0.8]
+
+* added gif support as network, memory and asset.
+
+## [1.0.9]
+
+* added BoxBorder to all image expect oval.
+
+## [1.1.0]
+
+* added BoxBorder to all image expect oval.
+
+## [1.2.0]
+
+* added BoxBorder to all image expect oval.
+
+## [1.3.0]
+
+* added support for animated json image [JsonAsset, JsonFile, JsonNetwork, JsonMemory].
+
+## [1.4.0]
+
+* added support for animated json image [JsonAsset, JsonFile, JsonNetwork, JsonMemory].
+
+## [1.5.0]
+
+* added support for animated json image [JsonAsset, JsonFile, JsonNetwork, JsonMemory].
+
+## [1.6.0]
+
+* update dependency.